requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
cli-exit-tools==1.2.3.2
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==4.4.2
dill==0.3.4
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lib-detect-testenv==2.0.3
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
moviepy==1.0.3
multiprocess==0.70.12.2
numpy==1.24.1
nvidia-cuda-runtime-cu12==12.0.107
nvidia-pyindex==1.0.9
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
proglog==0.1.10
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-version==0.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
wrapt-timeout-decorator==1.3.12.2
